GatheringDetail 컴포넌트 분할
