+++ src/components/Gathering/GatheringDetail.tsx
@@ -2,44 +2,21 @@ import React, { useState, useEffect, useRef } from "react";
 /** @jsxImportSource @emotion/react */ 
 import {css} from '@emotion/react';
 import {GatheringInfo} from '../../util/interfaces';
-import {divDate} from '../../util/util';
-import {GoThumbsup, GoThumbsdown} from 'react-icons/go';
-import {FiUpload, FiPocket} from 'react-icons/fi';
+import {divDate, setIntersectionObserver} from '../../util/util';
+import {GatheringFloatingBox} from '../index';
 
 interface Props {
     info: GatheringInfo | null
-    onBack: ()=> void
-    onJoin: (arg: React.MouseEvent<HTMLDivElement>)=> void
 }
 
-function GatheringDetail({info, onBack, onJoin}: Props){
+function GatheringDetail({info}: Props){
     const divRef = useRef<HTMLDivElement | null>(null);
-    const [float, setFloat] = useState(false);
-
+    const headerRef = useRef<HTMLDivElement | null>(null);
+    const [isFloat, setIsFloat] = useState(false);
     const nextAct = divDate(info?.nextActiveDate || '');
-    const initDate = divDate(info?.initDate || '');
-    const lastActiveDate = divDate(info?.lastActiveDate || '');
-    const isOnce = info?.filter.includes('한 번 만남');
 
-    const onClickBack = ()=>{
-        onBack();
-    }
-    
     const onClickDate = (e: React.MouseEvent<HTMLDivElement>)=>{
         selectDate(e);
-    }
-    
-    const onClickJoin = (e: React.MouseEvent<HTMLDivElement>)=>{
-        onJoin(e);
-    }
-    
-    const onClickIcon = (e: React.MouseEvent<HTMLDivElement>)=>{
-        const {type} = e.currentTarget.dataset;
-        if(type === 'share'){
-    
-        }else{
-    
-        }
     }
     
     const selectDate = (e: React.MouseEvent<HTMLDivElement>)=>{
@@ -53,40 +30,36 @@ function GatheringDetail({info, onBack, onJoin}: Props){
         }
     }
 
-    const setFloatingPosition = (e: Event)=>{
-        e.stopPropagation();
-        const header = divRef.current?.querySelector('.simple-info .header');
-        if(header instanceof HTMLElement){
-            const headerTop = header.getBoundingClientRect().top;
-            setFloat(window.scrollY + 20 > headerTop + window.scrollY);
-        }
+    const onClickModify = (e: React.MouseEvent<HTMLButtonElement>)=>{
+        const {name} = e.currentTarget;
+
     }
 
-    const attachScrollEvent = ()=>{
-        if(!window.onscroll){
-            window.onscroll = setFloatingPosition;
-        }
-    }
+    const setFloating = ()=>{
+        setIntersectionObserver({
+            root: null,
+            dom: headerRef.current!,
+            threshold: 1,
+            intersectionCallback: (el: Element, isIntersecting: boolean)=>{
+                setIsFloat(!isIntersecting);
+            }
+        });
+    } 
 
-    const removeScrollEvent = ()=>{
-        window.onscroll = null;
-    }
-    
     useEffect(()=>{
-        attachScrollEvent();
-        return ()=> {removeScrollEvent()}
+        setFloating();
     }, []);
-
+    
     return (
         !info ? null :
-        <div css={style(float)} ref={divRef}>
+        <div css={style} ref={divRef}>
             <div className='wrapper'>
                 <section className='info-section'>
                     <div className='simple-info'>
                         <div className='interests'>
                             {info.interests}
                         </div>
-                        <div className='header'>
+                        <div className='header' ref={headerRef}>
                             <div className='pay'>{info.payYn === 'Y' ? '유료' : '일반'}</div>
                             <div className='desc'>{info.description}</div>
                             {info.closeSoon === 'Y' ? 
@@ -118,6 +91,7 @@ function GatheringDetail({info, onBack, onJoin}: Props){
                                 <div className='gray'>{`(여 팀원${info.girl} 남 팀원${2})`}</div>
                                 <div className='can-apply'>신청가능</div>
                                 <div className='gray'>{`여${(info.girl || 0) - (info.currGirl || 0)} 남${(info.boy || 0) - (info.currBoy || 0)} `}</div>
+                                <button className='modify-btn' onClick={onClickModify}>수정</button>
                             </div>
                         </div>
                         <div className='row'>
@@ -133,6 +107,7 @@ function GatheringDetail({info, onBack, onJoin}: Props){
                             <div className='val'>
                                 {`${nextAct.MM}월 ${nextAct.dd}일 ${nextAct.HH} : ${nextAct.mm} ${nextAct.ampm}`} 
                                 <div className='else'>외 {info.activeDateList.length - 1}</div>
+                                <button className='modify-btn' onClick={onClickModify}>수정</button>
                             </div>
                         </div>
                         <div className='row'>
@@ -183,61 +158,15 @@ function GatheringDetail({info, onBack, onJoin}: Props){
                     </div>
                 </section>
                 <section className='floating-section'>
-                    <div className='floating-box' id='detailFloating'>
-                        <div className='title'>팀소개</div>
-                        <div className='content'>
-                            <div className='interests'>{info.interests}</div>
-                            <div className='name'>{info.name}</div>
-                            <div className='line1'></div>
-                            <div className='row'>
-                                <div className='txt1'>{`${initDate.MM}월 ${initDate.dd}일 개설`}</div>
-                                <div className='txt2'>{`${lastActiveDate.MM}월 ${lastActiveDate.dd}일에 마지막으로 모인 팀 입니다.`}</div>
-                            </div>
-                            <div className='row'>
-                                <div className='txt1'>{info.participants}명이 팀에 참여했습니다.</div>
-                                <div className='txt2'>현재 팀원은 {info.currentTeam}명 입니다.</div>
-                            </div>
-                            <div className='row age'>
-                                <div className='txt1'>평균연령 {info.averageAge}세</div>
-                                <div className='txt2'>남 {info.boyRate}%  여 {info.boyRate}%</div>
-                            </div>
-                            <div className='line2'></div>
-                            <div className='likes'>
-                                <div className='good'>
-                                    <GoThumbsup size='20'/>
-                                    <div>{info.likes}</div>
-                                </div>
-                                <div className='bad'>
-                                    <GoThumbsdown size='20'/>
-                                    <div>{info.hates}</div>
-                                </div>
-                            </div>
-                            <div className='btn' onClick={onClickJoin}>이 팀에 합류하기</div>
-                        </div>
-                        {isOnce ? 
-                            <div className='once-box'>
-                                <div className='btn once' data-type='once' onClick={onClickJoin}>한 번 참여하기</div>
-                                <div className='txt'>• 한 번 만남이 가능한 게시물입니다.</div>
-                            </div>
-                        : null}
-                        <div className='icons'>
-                            <div className='icon' data-type='share' onClick={onClickIcon}>
-                                <FiUpload size='25'/>
-                                <div className='txt'>공유하기</div>
-                            </div>
-                            <div className='icon' data-type='pocket' onClick={onClickIcon}>
-                                <FiPocket size='25'/>
-                                <div className='txt'>주머니 담기</div>
-                            </div>
-                        </div>
-                    </div>
+                    <GatheringFloatingBox info={info} isFloat={isFloat}/>
                 </section>
+
             </div>
         </div>
     );
 }
 
-const style =(float: boolean)=> (css`
+const style = css`
     .back-btn {
         width: 70px;
         text-align: center;
@@ -248,6 +177,14 @@ const style =(float: boolean)=> (css`
         color: var(--color-main-text);
         font-weight: bold;
         margin: 10px 0;
+    }
+    .modify-btn {
+        padding: 4px 8px;
+        font-size: 16px;
+        color: black;
+        background-color: var(--color-yellow);
+        margin: 0 20px;
+        border: 1px solid black;
     }
     .wrapper {
         display: flex;
@@ -333,8 +270,11 @@ const style =(float: boolean)=> (css`
                     }
                     .val {
                         display: flex;
+                        align-items: center;
                         font-size: 16px;
                         .can-apply {
+                            display: flex;
+                            align-items: center;
                             color: var(--color-peach);
                             border: 1px solid var(--color-peach);
                             border-radius: 10px;
@@ -431,115 +371,8 @@ const style =(float: boolean)=> (css`
             position: relative;
             width: 290px;
             padding-left: 24px;
-            .floating-box {
-                position: ${float ? 'fixed' : 'relative'};
-                top: ${float ? '20px' : '40px'};
-                .title {
-                    font-weight: bold;
-                    font-size: 24px;
-                    margin-bottom: 16px;
-                }
-                .content {
-                    width: 265px;
-                    padding: 12px;
-                    border: 1px solid #C4C4C4;
-                    border-radius: 10px;
-                    background-color: white;
-                    margin-bottom: 12px;
-                    .interests {
-                        color: var(--color-gray);
-                        font-size: 16px;
-                        margin-bottom: 24px;
-                    }
-                    .name {
-                        font-weight: bold;
-                        font-size: 16px;
-                    }
-                    .row {
-                        margin-bottom: 24px;
-                        .txt1 {
-                            font-weight: 600;
-                            font-size: 16px;
-                        }
-                        .txt2 {
-                            font-size: 12px;
-                            color: var(--color-gray);
-                            margin-top: 5px;
-                        }
-                    }
-                    .line1 {
-                        transform: translateX(-12px);
-                        width: calc(100% + 24px);
-                        margin: 12px 0 24px 0;
-                        height: 1px;
-                        background-color: #C4C4C4;
-                    }
-                    .line2 {
-                        margin-bottom: 20px;
-                        height: 1px;
-                        background-color: #C4C4C4;
-                    }
-                    .likes {
-                        display: flex;
-                        justify-content: center;
-                        margin-bottom: 40px;
-                        color: var(--color-gray);
-                        font-size: 24px;
-                        > div {
-                            margin: 0 25px;
-                            display: flex;
-                            flex-direction: column;
-                            align-items: center;
-                        }
-                        .good {
-                            color: var(--color-main-text);
-                        }
-                    }
-                }
-                .icons {
-                    display: flex;
-                    justify-content: center;
-                    margin-top: 19px;
-                    .icon {
-                        display: flex;
-                        flex-direction: column;
-                        align-items: center;
-                        margin: 0 12px;
-                        cursor: pointer;
-                        .txt {
-                            font-size: 16px;
-                        }
-                    }
-                }
-                .btn {
-                    position: relative;
-                    left: 50%;
-                    transform: translateX(-50%);
-                    width: 240px;
-                    height: 56px;
-                    line-height: 54px;
-                    font-size: 24px;
-                    font-weight: 500;
-                    text-align: center;
-                    border: 1px solid black;
-                    border-radius: 25px;
-                    background-color: var(--color-yellow);
-                    cursor: pointer;
-                }
-                .btn.once {
-                    background-color: var(--color-peach);
-                }
-                .once-box {
-                    .txt {
-                        text-align: center;
-                        font-size: 8px;
-                        color: var(--color-gray);
-                        margin-top: 4px;
-                    }
-                }
-            }
         }
     }
-`);
+`;
 
 export default GatheringDetail;
